Target Tables/cloud_functions/gold_layer/requirements.txt
functions-framework==3.*
duckdb==1.3.2
google-cloud-secret-manager
pandas
numpy
